+++ backends/rapidpro/backend_test.go
@@ -17,6 +17,7 @@ import (
 
 	"github.com/aws/aws-sdk-go/service/s3"
 	"github.com/aws/aws-sdk-go/service/s3/s3iface"
+	"github.com/garyburd/redigo/redis"
 	"github.com/nyaruka/courier"
 	"github.com/nyaruka/courier/queue"
 	"github.com/nyaruka/gocommon/urns"
@@ -797,6 +798,29 @@ func (ts *BackendTestSuite) TestWriteMsg() {
 	msg = ts.b.NewIncomingMsg(knChannel, urn, text).(*DBMsg)
 	err = writeMsgToDB(ctx, ts.b, msg)
 	ts.NoError(err)
+
+	// check that our mailroom queue has an item
+	rc := ts.b.redisPool.Get()
+	defer rc.Close()
+	rc.Do("DEL", "handler:1", "handler:active", fmt.Sprintf("c:1:%d", msg.ContactID_.Int64))
+
+	// test queuing to mailroom
+	knChannel.OrgFlowServerEnabled_ = true
+	msg = ts.b.NewIncomingMsg(knChannel, urn, "hello 1 2 3").(*DBMsg)
+	err = writeMsgToDB(ctx, ts.b, msg)
+	ts.NoError(err)
+
+	count, err := redis.Int(rc.Do("ZCARD", "handler:1"))
+	ts.NoError(err)
+	ts.Equal(1, count)
+
+	count, err = redis.Int(rc.Do("ZCARD", "handler:active"))
+	ts.NoError(err)
+	ts.Equal(1, count)
+
+	count, err = redis.Int(rc.Do("LLEN", fmt.Sprintf("c:1:%d", msg.ContactID_.Int64)))
+	ts.NoError(err)
+	ts.Equal(1, count)
 }
 
 func (ts *BackendTestSuite) TestChannelEvent() {
@@ -819,6 +843,45 @@ func (ts *BackendTestSuite) TestChannelEvent() {
 	ts.Equal(map[string]interface{}{"ref_id": "12345"}, dbE.Extra_.Map)
 	ts.Equal(contact.ID_, dbE.ContactID_)
 	ts.Equal(contact.URNID_, dbE.ContactURNID_)
+}
+
+func (ts *BackendTestSuite) TestMailroomEvents() {
+	ctx := context.Background()
+
+	rc := ts.b.redisPool.Get()
+	defer rc.Close()
+	rc.Do("DEL", "handler:1", "handler:active")
+
+	channel := ts.getChannel("KN", "dbc126ed-66bc-4e28-b67b-81dc3327c95d")
+	channel.OrgFlowServerEnabled_ = true
+	urn, _ := urns.NewTelURNForCountry("12065551616", channel.Country())
+	event := ts.b.NewChannelEvent(channel, courier.Referral, urn).WithExtra(map[string]interface{}{"ref_id": "12345"}).WithContactName("kermit frog")
+	err := ts.b.WriteChannelEvent(ctx, event)
+	ts.NoError(err)
+
+	contact, err := contactForURN(ctx, ts.b, channel.OrgID_, channel, urn, "", "")
+	ts.NoError(err)
+	ts.Equal("kermit frog", contact.Name_.String)
+
+	dbE := event.(*DBChannelEvent)
+	dbE, err = readChannelEventFromDB(ts.b, dbE.ID_)
+	ts.NoError(err)
+	ts.Equal(dbE.EventType_, courier.Referral)
+	ts.Equal(map[string]interface{}{"ref_id": "12345"}, dbE.Extra_.Map)
+	ts.Equal(contact.ID_, dbE.ContactID_)
+	ts.Equal(contact.URNID_, dbE.ContactURNID_)
+
+	count, err := redis.Int(rc.Do("ZCARD", "handler:1"))
+	ts.NoError(err)
+	ts.Equal(1, count)
+
+	count, err = redis.Int(rc.Do("ZCARD", "handler:active"))
+	ts.NoError(err)
+	ts.Equal(1, count)
+
+	count, err = redis.Int(rc.Do("LLEN", fmt.Sprintf("c:1:%d", contact.ID_.Int64)))
+	ts.NoError(err)
+	ts.Equal(1, count)
 }
 
 func TestMsgSuite(t *testing.T) {